Initial Commit
This commit adds a basic node struct, which will be used for each of the nodes in the system. It also adds a controller struct, which stores each node created in a Vec, and this commit includes a function to create a new node and add it to the vector.