Update Controller to use HashMap over Vec
I have changed the Controller to use a HashMap for storing all the nodes. This is so each node can be stored by a key: their ID, which makes finding the node in the map much easier, especially if I destroy nodes during runtime. Currently, I need to improve the get_node function so that it can handle if the ID given is not in the hashmap.