Change controller to spawn a new thread for each new node
This commit changes the system so that, when a new node is created, it spawns a new thread for that node. This is alongside the preexisting listener thread for each node. This means that the controller can't directly influence the nodes anymore, so I have given the controller its own udpsocket, which it will use to communicate with the other nodes and issue commands. A messages module has been created to prefix each message sent between controller and nodes, as well as node to node, with a specific message code.

The controller and node communication may be updated to TCP in a later implementation if it is decided that udp is too unreliable for commands. There is also currently no verification that a command has come from the controller, however given that the controller is purely for simulation I am not sure whether that is relevant for the security of the system.